requirements.txt
# Kalshi API
kalshi-python>=2.1.4

# Config
python-dotenv>=1.0.0

# LLM (optional - for narrative summarization)
anthropic>=0.40.0

# Sentiment analysis
transformers>=4.40.0
torch>=2.2.0
vaderSentiment>=3.3.2
sentence-transformers>=2.5.0

# Data sources
praw>=7.7.0
newsapi-python>=0.2.7
feedparser>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# ML / Statistics
xgboost>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0

# Async
aiohttp>=3.9.0
